implemented changes with proxy and refactored actions
