+++ src/Pages/Itineraries/Itineraries.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import BackButton from "../../components/BackButton/BackButton";
 import Itinerary from "../../components/Itinerary/Itinerary";
 import { connect } from "react-redux";
-import { fetchIterinariesForOneCity } from "../../actions";
+import { fetchIterinariesForOneCity } from "../../actions/itinerariesActions";
 import "./Itineraries.css";
 
 class Itineraries extends Component {
