+++ src/actions/types.js
@@ -1,4 +1,4 @@
-export const FETCH_CITY = "FETCH_CITY";
+export const FETCH_CITIES = "FETCH_CITIES";
 
 export const FETCH_ITINERARIES = "FETCH_ININERARIES";
