+++ src/components/Menu/Menu.js
@@ -4,10 +4,10 @@ import "./Menu.css";
 
 class Menu extends Component {
   render() {
-    // Render nothing if the "show" prop is false
-    if (!this.props.show) {
-      return null;
-    }
+    // // Render nothing if the "show" prop is false
+    // if (!this.props.show) {
+    //   return null;
+    // }
 
     return (
       <div className="outerWrapper">
