+++ src/reducers/cityReducer.js
@@ -1,8 +1,8 @@
-import { FETCH_CITY } from "../actions/types";
+import { FETCH_CITIES } from "../actions/types";
 
 export default function cityReducer(state = [], action) {
   switch (action.type) {
-    case FETCH_CITY:
+    case FETCH_CITIES:
       return action.cities;
     default:
       return state;
